fix(folder structure): move a hook to right folder
